working simple user rest

+++ src/main/java/com/epam/turik/restchat/rest/UserController.java
@@ -2,7 +2,10 @@ package com.epam.turik.restchat.rest;
 
 import com.epam.turik.restchat.data.UserRepository;
 import com.epam.turik.restchat.data.objects.User;
+import com.epam.turik.restchat.rest.exceptions.UserNotFoundException;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
 
 @RestController
 @RequestMapping("user")
@@ -15,7 +18,15 @@ public class UserController {
 
     @GetMapping("/{id}")
     public User getUser(@PathVariable Long id) {
+        if (!userRepository.existsById(id)) {
+            throw new UserNotFoundException();
+        }
         return userRepository.findById(id).orElse(null);
+    }
+
+    @GetMapping("/all")
+    public List<User> getAll() {
+        return (List<User>) userRepository.findAll();
     }
 
     @PostMapping("/")
@@ -23,13 +34,21 @@ public class UserController {
         userRepository.save(user);
     }
 
-    @DeleteMapping("/id")
+    @DeleteMapping("/{id}")
     public void deleteUser(@PathVariable Long id) {
+        if (!userRepository.existsById(id)) {
+            throw new UserNotFoundException();
+        }
         userRepository.deleteById(id);
     }
 
-    @PutMapping("/")
-    public void updateUser(@RequestBody User user){
-        System.out.println(user);
+    @PutMapping("/{id}")
+    public void updateUser(@PathVariable Long id, @RequestBody User data){
+        if (!userRepository.existsById(id)) {
+            throw new UserNotFoundException();
+        }
+        User user = userRepository.findById(id).orElse(null);
+        user.setUsername(data.getUsername());
+        userRepository.save(user);
     }
 }
